# GH-799 Improve dependency loader performance with parallel downloading and relocation.