--- a/buildSrc/src/main/kotlin/eternalcore-repositories.gradle.kts
+++ b/buildSrc/src/main/kotlin/eternalcore-repositories.gradle.kts
@@ -5,11 +5,8 @@ plugins {
 repositories {
     mavenCentral()
     maven { url = uri("https://hub.spigotmc.org/nexus/content/repositories/snapshots/") }
-    maven { url = uri("https://repo.panda-lang.org/releases/") }
     maven { url = uri("https://repo.papermc.io/repository/maven-public/") }
-    maven { url = uri("https://repository.minecodes.pl/releases/") }
-    maven { url = uri("https://repository.minecodes.pl/snapshots/") }
-    maven { url = uri("https://repo.eternalcode.pl/snapshots/") }
+    maven { url = uri("https://repo.panda-lang.org/releases/") }
     maven { url = uri("https://repo.eternalcode.pl/releases/") }
     maven { url = uri("https://repo.extendedclip.com/content/repositories/placeholderapi/") }
 }