--- a/eternalcore-plugin/src/main/java/com/eternalcode/core/loader/dependency/DependencyLoadEntry.java
+++ b/eternalcore-plugin/src/main/java/com/eternalcode/core/loader/dependency/DependencyLoadEntry.java
@@ -1,0 +1,6 @@
+package com.eternalcode.core.loader.dependency;
+
+import java.nio.file.Path;
+
+public record DependencyLoadEntry(Dependency dependency, Path path) {
+}