--- a/eternalcore-plugin/src/main/java/com/eternalcode/core/loader/dependency/DependencyLoadResult.java
+++ b/eternalcore-plugin/src/main/java/com/eternalcode/core/loader/dependency/DependencyLoadResult.java
@@ -2,9 +2,7 @@ package com.eternalcode.core.loader.dependency;
 
 import com.eternalcode.core.loader.classloader.IsolatedClassLoader;
 
-import java.nio.file.Path;
-import java.util.Collection;
 import java.util.List;
 
-public record DependencyLoadResult(IsolatedClassLoader loader, Collection<Dependency> dependencies, List<Path> paths) {
+public record DependencyLoadResult(IsolatedClassLoader loader, List<DependencyLoadEntry> paths) {
 }