--- a/eternalcore-plugin/src/main/java/com/eternalcode/core/loader/pom/PomXmlScanner.java
+++ b/eternalcore-plugin/src/main/java/com/eternalcode/core/loader/pom/PomXmlScanner.java
@@ -68,7 +68,7 @@ public class PomXmlScanner implements DependencyScanner {
                 }
 
                 collector = this.findAllChildren(collector, firstChild);
-                collector.scannedDependency(firstChild);
+                collector.addScannedDependency(firstChild);
             }
 
             break;