Resolve: #799

# Tests
A - Rollczi's computer 6x 4.90GHz, 48GB, 40Mbps
B - vLucky's computer 12Mbps
C - VPS 4x 3,7 GHz, 8GB, 730Mbps

### Before
A - 90s (38s + 52s)
B - 98s (45s + 53s)
C - 54s (19s + 35s) 

### After
A - 21s (15s + 6s)
B - 31s (18s + 13s)
C - 12s (10s + 2s)

### Results
A - **4,2x** faster
B - **3,2x** faster
C - **4,5x** faster

(first part of the time represents searching for dependency, second represents downloading and relocation)